move assets to separate script and re-use with storybook
