+++ web/scripts/build-web.mjs
@@ -5,8 +5,9 @@
 import "@goauthentik/core/environment/load/node";
 
 import * as fs from "node:fs/promises";
-import { createRequire } from "node:module";
 import * as path from "node:path";
+
+import { copyAssets } from "./build-assets.mjs";
 
 /**
  * @file ESBuild script for building the authentik web UI.
@@ -27,8 +28,6 @@ import esbuild from "esbuild";
 
 /// <reference types="../types/esbuild.js" />
 
-const require = createRequire(import.meta.url);
-
 const logger = ConsoleLogger.child({ name: "Build" });
 
 const bundleDefinitions = createBundleDefinitions();
@@ -37,49 +36,6 @@ const publicBundledDefinitions = Object.fromEntries(
     Object.entries(bundleDefinitions).map(([name, value]) => [name, JSON.parse(value)]),
 );
 logger.info(publicBundledDefinitions, "Bundle definitions");
-
-/**
- * @typedef {[from: string, to: string]} SourceDestinationPair
- */
-
-/**
- * @type {SourceDestinationPair[]}
- */
-const assets = [
-    [
-        path.join(path.dirname(EntryPoint.StandaloneLoading.in), "startup"),
-        path.dirname(EntryPoint.StandaloneLoading.out),
-    ],
-    [path.resolve(PackageRoot, "src", "assets", "images"), "./assets/images"],
-    [require.resolve("@goauthentik/brand-assets/brand.png"), "./assets/icons/brand.png"],
-    [require.resolve("@goauthentik/brand-assets/brand.svg"), "./assets/icons/brand.svg"],
-    [require.resolve("@goauthentik/brand-assets/icon.png"), "./assets/icons/icon.png"],
-    [require.resolve("@goauthentik/brand-assets/icon.svg"), "./assets/icons/icon.svg"],
-    [
-        require.resolve("@goauthentik/brand-assets/icon_left_brand.png"),
-        "./assets/icons/icon_left_brand.png",
-    ],
-    [
-        require.resolve("@goauthentik/brand-assets/icon_left_brand.svg"),
-        "./assets/icons/icon_left_brand.svg",
-    ],
-    [
-        require.resolve("@goauthentik/brand-assets/icon_pride_lgbt.png"),
-        "./assets/icons/icon_pride_lgbt.png",
-    ],
-    [
-        require.resolve("@goauthentik/brand-assets/icon_pride_trans.png"),
-        "./assets/icons/icon_pride_trans.png",
-    ],
-    [
-        require.resolve("@goauthentik/brand-assets/icon_top_brand.png"),
-        "./assets/icons/icon_top_brand.png",
-    ],
-    [
-        require.resolve("@goauthentik/brand-assets/icon_top_brand.svg"),
-        "./assets/icons/icon_top_brand.svg",
-    ],
-];
 
 const entryPointNames = Object.keys(EntryPoint);
 const entryPoints = Object.values(EntryPoint);
@@ -98,29 +54,7 @@ const BASE_ESBUILD_PLUGINS = [
                  */
                 const errors = [];
 
-                /**
-                 * @param {SourceDestinationPair} pair
-                 */
-                const copy = ([from, to]) => {
-                    const resolvedDestination = path.resolve(DistDirectory, to);
-
-                    logger.debug(`📋 Copying assets from ${from} to ${to}`);
-
-                    return fs
-                        .cp(from, resolvedDestination, {
-                            recursive: true,
-                        })
-                        .catch((error) => {
-                            errors.push({
-                                text: `Failed to copy assets from ${from} to ${to}: ${error}`,
-                                location: {
-                                    file: from,
-                                },
-                            });
-                        });
-                };
-
-                await Promise.all(assets.map(copy));
+                await copyAssets();
 
                 return { errors };
             });
